moving all creation logic to separate file to keep the api layer simple
